// src/lib/database.types.ts
export type Size = {
  id: number
  code: string
  description: string | null
  created_at: string
}
// ...
export type PriceHistory = {
  id: number
  product_variant_id: number
  old_wholesale_price: number | null
  new_wholesale_price: number | null
  old_retail_price: number | null
  new_retail_price: number | null
  changed_at: string
}